+++ .travis/deploy_docs.sh
@@ -4,7 +4,7 @@
 
 set -o errexit -o nounset
 
-if [ "$TRAVIS_REPO_SLUG" != "ofek/bit" ] || [ "$TRAVIS_PULL_REQUEST" != "false" ] || [ "$TRAVIS_BRANCH" != "master" ]
+if [ "$TRAVIS_REPO_SLUG" != "ofek/bitcash" ] || [ "$TRAVIS_PULL_REQUEST" != "false" ] || [ "$TRAVIS_BRANCH" != "master" ]
 then
   echo "This commit was made against the $TRAVIS_BRANCH and not the master! No deploy!"
   exit 0
@@ -32,7 +32,7 @@ git init
 git config user.name "Ofek Lev"
 git config user.email "ofekmeister@gmail.com"
 
-git remote add upstream "https://$GH_TOKEN@github.com/ofek/bit.git"
+git remote add upstream "https://$GH_TOKEN@github.com/ofek/bitcash.git"
 git fetch upstream
 git reset upstream/gh-pages