+++ bitcash/network/fees.py
@@ -7,7 +7,7 @@ from requests.exceptions import ConnectionError, HTTPError, Timeout
 DEFAULT_FEE_FAST = 220
 DEFAULT_FEE_HOUR = 160
 DEFAULT_CACHE_TIME = 60 * 10
-URL = 'https://bitcoinfees.earn.com/api/v1/fees/recommended'
+URL = 'https://bitcoincashfees.earn.com/api/v1/fees/recommended'
 
 
 def set_fee_cache_time(seconds):