+++ bitcash/wallet.py
@@ -1,13 +1,13 @@
 import json
 
-from bit.crypto import ECPrivateKey
-from bit.curve import Point
-from bit.format import (
+from bitcash.crypto import ECPrivateKey
+from bitcash.curve import Point
+from bitcash.format import (
     bytes_to_wif, public_key_to_address, public_key_to_coords, wif_to_bytes
 )
-from bit.network import NetworkAPI, get_fee_cached, satoshi_to_currency_cached
-from bit.network.meta import Unspent
-from bit.transaction import calc_txid, create_p2pkh_transaction, sanitize_tx_data
+from bitcash.network import NetworkAPI, get_fee_cached, satoshi_to_currency_cached
+from bitcash.network.meta import Unspent
+from bitcash.transaction import calc_txid, create_p2pkh_transaction, sanitize_tx_data
 
 
 def wif_to_key(wif):
@@ -122,7 +122,7 @@ class BaseKey:
 
 
 class PrivateKey(BaseKey):
-    """This class represents a Bitcoin private key. ``Key`` is an alias.
+    """This class represents a BitcoinCash private key. ``Key`` is an alias.
 
     :param wif: A private key serialized to the Wallet Import Format. If the
                 argument is not supplied, a new private key will be created.
@@ -166,8 +166,8 @@ class PrivateKey(BaseKey):
 
     def get_balance(self, currency='satoshi'):
         """Fetches the current balance by calling
-        :func:`~bit.PrivateKey.get_unspents` and returns it using
-        :func:`~bit.PrivateKey.balance_as`.
+        :func:`~bitcash.PrivateKey.get_unspents` and returns it using
+        :func:`~bitcash.PrivateKey.balance_as`.
 
         :param currency: One of the :ref:`supported currencies`.
         :type currency: ``str``
@@ -179,7 +179,7 @@ class PrivateKey(BaseKey):
     def get_unspents(self):
         """Fetches all available unspent transaction outputs.
 
-        :rtype: ``list`` of :class:`~bit.network.meta.Unspent`
+        :rtype: ``list`` of :class:`~bitcash.network.meta.Unspent`
         """
         self.unspents[:] = NetworkAPI.get_unspent(self.address)
         self.balance = sum(unspent.amount for unspent in self.unspents)
@@ -204,25 +204,25 @@ class PrivateKey(BaseKey):
                         must be :ref:`supported <supported currencies>`.
         :type outputs: ``list`` of ``tuple``
         :param fee: The number of satoshi per byte to pay to miners. By default
-                    Bit will poll `<https://bitcoinfees.earn.com>`_ and use a fee
+                    Bitcash will poll `<https://bitcoincashfees.earn.com>`_ and use a fee
                     that will allow your transaction to be confirmed as soon as
                     possible.
         :type fee: ``int``
         :param leftover: The destination that will receive any change from the
-                         transaction. By default Bit will send any change to
+                         transaction. By default Bitcash will send any change to
                          the same address you sent from.
         :type leftover: ``str``
-        :param combine: Whether or not Bit should use all available UTXOs to
+        :param combine: Whether or not Bitcash should use all available UTXOs to
                         make future transactions smaller and therefore reduce
-                        fees. By default Bit will consolidate UTXOs.
+                        fees. By default Bitcash will consolidate UTXOs.
         :type combine: ``bool``
         :param message: A message to include in the transaction. This will be
                         stored in the blockchain forever. Due to size limits,
                         each message will be stored in chunks of 40 bytes.
         :type message: ``str``
-        :param unspents: The UTXOs to use as the inputs. By default Bit will
+        :param unspents: The UTXOs to use as the inputs. By default Bitcash will
                          communicate with the blockchain itself.
-        :type unspents: ``list`` of :class:`~bit.network.meta.Unspent`
+        :type unspents: ``list`` of :class:`~bitcash.network.meta.Unspent`
         :returns: The signed transaction as hex.
         :rtype: ``str``
         """
@@ -243,7 +243,7 @@ class PrivateKey(BaseKey):
              message=None, unspents=None):  # pragma: no cover
         """Creates a signed P2PKH transaction and attempts to broadcast it on
         the blockchain. This accepts the same arguments as
-        :func:`~bit.PrivateKey.create_transaction`.
+        :func:`~bitcash.PrivateKey.create_transaction`.
 
         :param outputs: A sequence of outputs you wish to send in the form
                         ``(destination, amount, currency)``. The amount can
@@ -252,25 +252,25 @@ class PrivateKey(BaseKey):
                         must be :ref:`supported <supported currencies>`.
         :type outputs: ``list`` of ``tuple``
         :param fee: The number of satoshi per byte to pay to miners. By default
-                    Bit will poll `<https://bitcoinfees.earn.com>`_ and use a fee
+                    Bitcash will poll `<https://bitcoincashfees.earn.com>`_ and use a fee
                     that will allow your transaction to be confirmed as soon as
                     possible.
         :type fee: ``int``
         :param leftover: The destination that will receive any change from the
-                         transaction. By default Bit will send any change to
+                         transaction. By default Bitcash will send any change to
                          the same address you sent from.
         :type leftover: ``str``
-        :param combine: Whether or not Bit should use all available UTXOs to
+        :param combine: Whether or not Bitcash should use all available UTXOs to
                         make future transactions smaller and therefore reduce
-                        fees. By default Bit will consolidate UTXOs.
+                        fees. By default Bitcash will consolidate UTXOs.
         :type combine: ``bool``
         :param message: A message to include in the transaction. This will be
                         stored in the blockchain forever. Due to size limits,
                         each message will be stored in chunks of 40 bytes.
         :type message: ``str``
-        :param unspents: The UTXOs to use as the inputs. By default Bit will
+        :param unspents: The UTXOs to use as the inputs. By default Bitcash will
                          communicate with the blockchain itself.
-        :type unspents: ``list`` of :class:`~bit.network.meta.Unspent`
+        :type unspents: ``list`` of :class:`~bitcash.network.meta.Unspent`
         :returns: The transaction ID.
         :rtype: ``str``
         """
@@ -300,25 +300,25 @@ class PrivateKey(BaseKey):
                            compressed public key. This influences the fee.
         :type compressed: ``bool``
         :param fee: The number of satoshi per byte to pay to miners. By default
-                    Bit will poll `<https://bitcoinfees.earn.com>`_ and use a fee
+                    Bitcash will poll `<https://bitcoincashfees.earn.com>`_ and use a fee
                     that will allow your transaction to be confirmed as soon as
                     possible.
         :type fee: ``int``
         :param leftover: The destination that will receive any change from the
-                         transaction. By default Bit will send any change to
+                         transaction. By default Bitcash will send any change to
                          the same address you sent from.
         :type leftover: ``str``
-        :param combine: Whether or not Bit should use all available UTXOs to
+        :param combine: Whether or not Bitcash should use all available UTXOs to
                         make future transactions smaller and therefore reduce
-                        fees. By default Bit will consolidate UTXOs.
+                        fees. By default Bitcash will consolidate UTXOs.
         :type combine: ``bool``
         :param message: A message to include in the transaction. This will be
                         stored in the blockchain forever. Due to size limits,
                         each message will be stored in chunks of 40 bytes.
         :type message: ``str``
-        :param unspents: The UTXOs to use as the inputs. By default Bit will
+        :param unspents: The UTXOs to use as the inputs. By default Bitcash will
                          communicate with the blockchain itself.
-        :type unspents: ``list`` of :class:`~bit.network.meta.Unspent`
+        :type unspents: ``list`` of :class:`~bitcash.network.meta.Unspent`
         :returns: JSON storing data required to create an offline transaction.
         :rtype: ``str``
         """
@@ -343,7 +343,7 @@ class PrivateKey(BaseKey):
         """Creates a signed P2PKH transaction using previously prepared
         transaction data.
 
-        :param tx_data: Output of :func:`~bit.PrivateKey.prepare_transaction`.
+        :param tx_data: Output of :func:`~bitcash.PrivateKey.prepare_transaction`.
         :type tx_data: ``str``
         :returns: The signed transaction as hex.
         :rtype: ``str``
@@ -360,7 +360,7 @@ class PrivateKey(BaseKey):
         """
         :param hexed: A private key previously encoded as hex.
         :type hexed: ``str``
-        :rtype: :class:`~bit.PrivateKey`
+        :rtype: :class:`~bitcash.PrivateKey`
         """
         return PrivateKey(ECPrivateKey.from_hex(hexed))
 
@@ -369,7 +369,7 @@ class PrivateKey(BaseKey):
         """
         :param bytestr: A private key previously encoded as hex.
         :type bytestr: ``bytes``
-        :rtype: :class:`~bit.PrivateKey`
+        :rtype: :class:`~bitcash.PrivateKey`
         """
         return PrivateKey(ECPrivateKey(bytestr))
 
@@ -378,7 +378,7 @@ class PrivateKey(BaseKey):
         """
         :param der: A private key previously encoded as DER.
         :type der: ``bytes``
-        :rtype: :class:`~bit.PrivateKey`
+        :rtype: :class:`~bitcash.PrivateKey`
         """
         return PrivateKey(ECPrivateKey.from_der(der))
 
@@ -387,7 +387,7 @@ class PrivateKey(BaseKey):
         """
         :param pem: A private key previously encoded as PEM.
         :type pem: ``bytes``
-        :rtype: :class:`~bit.PrivateKey`
+        :rtype: :class:`~bitcash.PrivateKey`
         """
         return PrivateKey(ECPrivateKey.from_pem(pem))
 
@@ -396,7 +396,7 @@ class PrivateKey(BaseKey):
         """
         :param num: A private key in raw integer form.
         :type num: ``int``
-        :rtype: :class:`~bit.PrivateKey`
+        :rtype: :class:`~bitcash.PrivateKey`
         """
         return PrivateKey(ECPrivateKey.from_int(num))
 
@@ -405,7 +405,7 @@ class PrivateKey(BaseKey):
 
 
 class PrivateKeyTestnet(BaseKey):
-    """This class represents a testnet Bitcoin private key. **Note:** coins
+    """This class represents a testnet BitcoinCash private key. **Note:** coins
     on the test network have no monetary value!
 
     :param wif: A private key serialized to the Wallet Import Format. If the
@@ -450,8 +450,8 @@ class PrivateKeyTestnet(BaseKey):
 
     def get_balance(self, currency='satoshi'):
         """Fetches the current balance by calling
-        :func:`~bit.PrivateKeyTestnet.get_unspents` and returns it using
-        :func:`~bit.PrivateKeyTestnet.balance_as`.
+        :func:`~bitcash.PrivateKeyTestnet.get_unspents` and returns it using
+        :func:`~bitcash.PrivateKeyTestnet.balance_as`.
 
         :param currency: One of the :ref:`supported currencies`.
         :type currency: ``str``
@@ -463,7 +463,7 @@ class PrivateKeyTestnet(BaseKey):
     def get_unspents(self):
         """Fetches all available unspent transaction outputs.
 
-        :rtype: ``list`` of :class:`~bit.network.meta.Unspent`
+        :rtype: ``list`` of :class:`~bitcash.network.meta.Unspent`
         """
         self.unspents[:] = NetworkAPI.get_unspent_testnet(self.address)
         self.balance = sum(unspent.amount for unspent in self.unspents)
@@ -488,25 +488,25 @@ class PrivateKeyTestnet(BaseKey):
                         must be :ref:`supported <supported currencies>`.
         :type outputs: ``list`` of ``tuple``
         :param fee: The number of satoshi per byte to pay to miners. By default
-                    Bit will poll `<https://bitcoinfees.earn.com>`_ and use a fee
+                    Bitcash will poll `<https://bitcoincashfees.earn.com>`_ and use a fee
                     that will allow your transaction to be confirmed as soon as
                     possible.
         :type fee: ``int``
         :param leftover: The destination that will receive any change from the
-                         transaction. By default Bit will send any change to
+                         transaction. By default Bitcash will send any change to
                          the same address you sent from.
         :type leftover: ``str``
-        :param combine: Whether or not Bit should use all available UTXOs to
+        :param combine: Whether or not Bitcash should use all available UTXOs to
                         make future transactions smaller and therefore reduce
-                        fees. By default Bit will consolidate UTXOs.
+                        fees. By default Bitcash will consolidate UTXOs.
         :type combine: ``bool``
         :param message: A message to include in the transaction. This will be
                         stored in the blockchain forever. Due to size limits,
                         each message will be stored in chunks of 40 bytes.
         :type message: ``str``
-        :param unspents: The UTXOs to use as the inputs. By default Bit will
+        :param unspents: The UTXOs to use as the inputs. By default Bitcash will
                          communicate with the testnet blockchain itself.
-        :type unspents: ``list`` of :class:`~bit.network.meta.Unspent`
+        :type unspents: ``list`` of :class:`~bitcash.network.meta.Unspent`
         :returns: The signed transaction as hex.
         :rtype: ``str``
         """
@@ -527,7 +527,7 @@ class PrivateKeyTestnet(BaseKey):
              message=None, unspents=None):
         """Creates a signed P2PKH transaction and attempts to broadcast it on
         the testnet blockchain. This accepts the same arguments as
-        :func:`~bit.PrivateKeyTestnet.create_transaction`.
+        :func:`~bitcash.PrivateKeyTestnet.create_transaction`.
 
         :param outputs: A sequence of outputs you wish to send in the form
                         ``(destination, amount, currency)``. The amount can
@@ -536,25 +536,25 @@ class PrivateKeyTestnet(BaseKey):
                         must be :ref:`supported <supported currencies>`.
         :type outputs: ``list`` of ``tuple``
         :param fee: The number of satoshi per byte to pay to miners. By default
-                    Bit will poll `<https://bitcoinfees.earn.com>`_ and use a fee
+                    Bitcash will poll `<https://bitcoincashfees.earn.com>`_ and use a fee
                     that will allow your transaction to be confirmed as soon as
                     possible.
         :type fee: ``int``
         :param leftover: The destination that will receive any change from the
-                         transaction. By default Bit will send any change to
+                         transaction. By default Bitcash will send any change to
                          the same address you sent from.
         :type leftover: ``str``
-        :param combine: Whether or not Bit should use all available UTXOs to
+        :param combine: Whether or not Bitcash should use all available UTXOs to
                         make future transactions smaller and therefore reduce
-                        fees. By default Bit will consolidate UTXOs.
+                        fees. By default Bitcash will consolidate UTXOs.
         :type combine: ``bool``
         :param message: A message to include in the transaction. This will be
                         stored in the blockchain forever. Due to size limits,
                         each message will be stored in chunks of 40 bytes.
         :type message: ``str``
-        :param unspents: The UTXOs to use as the inputs. By default Bit will
+        :param unspents: The UTXOs to use as the inputs. By default Bitcash will
                          communicate with the testnet blockchain itself.
-        :type unspents: ``list`` of :class:`~bit.network.meta.Unspent`
+        :type unspents: ``list`` of :class:`~bitcash.network.meta.Unspent`
         :returns: The transaction ID.
         :rtype: ``str``
         """
@@ -584,25 +584,25 @@ class PrivateKeyTestnet(BaseKey):
                            compressed public key. This influences the fee.
         :type compressed: ``bool``
         :param fee: The number of satoshi per byte to pay to miners. By default
-                    Bit will poll `<https://bitcoinfees.earn.com>`_ and use a fee
+                    Bitcash will poll `<https://bitcoincashfees.earn.com>`_ and use a fee
                     that will allow your transaction to be confirmed as soon as
                     possible.
         :type fee: ``int``
         :param leftover: The destination that will receive any change from the
-                         transaction. By default Bit will send any change to
+                         transaction. By default Bitcash will send any change to
                          the same address you sent from.
         :type leftover: ``str``
-        :param combine: Whether or not Bit should use all available UTXOs to
+        :param combine: Whether or not Bitcash should use all available UTXOs to
                         make future transactions smaller and therefore reduce
-                        fees. By default Bit will consolidate UTXOs.
+                        fees. By default Bitcash will consolidate UTXOs.
         :type combine: ``bool``
         :param message: A message to include in the transaction. This will be
                         stored in the blockchain forever. Due to size limits,
                         each message will be stored in chunks of 40 bytes.
         :type message: ``str``
-        :param unspents: The UTXOs to use as the inputs. By default Bit will
+        :param unspents: The UTXOs to use as the inputs. By default Bitcash will
                          communicate with the blockchain itself.
-        :type unspents: ``list`` of :class:`~bit.network.meta.Unspent`
+        :type unspents: ``list`` of :class:`~bitcash.network.meta.Unspent`
         :returns: JSON storing data required to create an offline transaction.
         :rtype: ``str``
         """
@@ -627,7 +627,7 @@ class PrivateKeyTestnet(BaseKey):
         """Creates a signed P2PKH transaction using previously prepared
         transaction data.
 
-        :param tx_data: Output of :func:`~bit.PrivateKeyTestnet.prepare_transaction`.
+        :param tx_data: Output of :func:`~bitcash.PrivateKeyTestnet.prepare_transaction`.
         :type tx_data: ``str``
         :returns: The signed transaction as hex.
         :rtype: ``str``
@@ -644,7 +644,7 @@ class PrivateKeyTestnet(BaseKey):
         """
         :param hexed: A private key previously encoded as hex.
         :type hexed: ``str``
-        :rtype: :class:`~bit.PrivateKeyTestnet`
+        :rtype: :class:`~bitcash.PrivateKeyTestnet`
         """
         return PrivateKeyTestnet(ECPrivateKey.from_hex(hexed))
 
@@ -653,7 +653,7 @@ class PrivateKeyTestnet(BaseKey):
         """
         :param bytestr: A private key previously encoded as hex.
         :type bytestr: ``bytes``
-        :rtype: :class:`~bit.PrivateKeyTestnet`
+        :rtype: :class:`~bitcash.PrivateKeyTestnet`
         """
         return PrivateKeyTestnet(ECPrivateKey(bytestr))
 
@@ -662,7 +662,7 @@ class PrivateKeyTestnet(BaseKey):
         """
         :param der: A private key previously encoded as DER.
         :type der: ``bytes``
-        :rtype: :class:`~bit.PrivateKeyTestnet`
+        :rtype: :class:`~bitcash.PrivateKeyTestnet`
         """
         return PrivateKeyTestnet(ECPrivateKey.from_der(der))
 
@@ -671,7 +671,7 @@ class PrivateKeyTestnet(BaseKey):
         """
         :param pem: A private key previously encoded as PEM.
         :type pem: ``bytes``
-        :rtype: :class:`~bit.PrivateKeyTestnet`
+        :rtype: :class:`~bitcash.PrivateKeyTestnet`
         """
         return PrivateKeyTestnet(ECPrivateKey.from_pem(pem))
 
@@ -680,7 +680,7 @@ class PrivateKeyTestnet(BaseKey):
         """
         :param num: A private key in raw integer form.
         :type num: ``int``
-        :rtype: :class:`~bit.PrivateKeyTestnet`
+        :rtype: :class:`~bitcash.PrivateKeyTestnet`
         """
         return PrivateKeyTestnet(ECPrivateKey.from_int(num))