+++ docs/source/conf.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# Bit documentation build configuration file, created by
+# Bitcash documentation build configuration file, created by
 # sphinx-quickstart on Mon Feb 20 15:41:44 2017.
 #
 # This file is execfile()d with the current directory set to its
@@ -21,7 +21,7 @@
 # import sys
 # sys.path.insert(0, os.path.abspath('.'))
 
-from bit import __version__
+from bitcash import __version__
 
 
 # -- General configuration ------------------------------------------------
@@ -52,7 +52,7 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = 'Bit'
+project = 'Bitcash'
 copyright = '2017, Ofek Lev'
 author = 'Ofek Lev'
 
@@ -79,7 +79,7 @@ html_sidebars = {
 html_theme_options = {
     'show_powered_by': False,
     'github_user': 'ofek',
-    'github_repo': 'bit',
+    'github_repo': 'bitcash',
     'github_banner': True,
     'show_related': False
 }
@@ -126,7 +126,7 @@ html_static_path = ['_static']
 # -- Options for HTMLHelp output ------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'Bitdoc'
+htmlhelp_basename = 'Bitcashdoc'
 
 
 # -- Options for LaTeX output ---------------------------------------------
@@ -153,7 +153,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'Bit.tex', 'Bit Documentation',
+    (master_doc, 'Bitcash.tex', 'Bitcash Documentation',
      'Ofek Lev', 'manual'),
 ]
 
@@ -163,7 +163,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'bit', 'Bit Documentation',
+    (master_doc, 'bitcash', 'Bitcash Documentation',
      [author], 1)
 ]
 
@@ -174,8 +174,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'Bit', 'Bit Documentation',
-     author, 'Bit', 'One line description of project.',
+    (master_doc, 'Bitcash', 'Bitcash Documentation',
+     author, 'Bitcash', 'One line description of project.',
      'Miscellaneous'),
 ]